Add template filter to pages list view

+++ Services/TaskService.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace RevisionTen\CMS\Services;
 
 use DateTime;
+use Exception;
+use RevisionTen\CMS\Command\PageRemoveScheduleCommand;
+use RevisionTen\CMS\Model\Page;
 use RevisionTen\CMS\Model\Task;
 use RevisionTen\CQRS\Model\EventStreamObject;
 use RevisionTen\CQRS\Services\CommandBus;
@@ -12,33 +15,14 @@ use Doctrine\ORM\EntityManagerInterface;
 use RevisionTen\CQRS\Services\MessageBus;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Class TaskService.
- */
 class TaskService
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    protected $em;
+    protected EntityManagerInterface $em;
 
-    /**
-     * @var CommandBus
-     */
-    protected $commandBus;
+    protected CommandBus $commandBus;
 
-    /**
-     * @var CommandBus
-     */
-    protected $messageBus;
+    protected MessageBus $messageBus;
 
-    /**
-     * TaskService constructor.
-     *
-     * @param EntityManagerInterface  $em
-     * @param CommandBus              $commandBus
-     * @param MessageBus              $messageBus
-     */
     public function __construct(EntityManagerInterface $em, CommandBus $commandBus, MessageBus $messageBus)
     {
         $this->em = $em;
@@ -62,7 +46,9 @@ class TaskService
 
     public function markTasksAsDeleted(string $aggregateUuid): void
     {
-        /** @var Task[] $tasks */
+        /**
+         * @var Task[] $tasks
+         */
         $tasks = $this->em->getRepository(Task::class)->findBy(['aggregateUuid' => $aggregateUuid]);
 
         foreach ($tasks as $task) {
@@ -76,11 +62,20 @@ class TaskService
 
     public function removeTask(string $taskUuid): void
     {
-        /** @var Task[] $tasks */
+        /**
+         * @var Task[] $tasks
+         */
         $tasks = $this->em->getRepository(Task::class)->findBy(['uuid' => $taskUuid]);
 
         foreach ($tasks as $task) {
-            $this->em->remove($task);
+            if (null !== $task->getResultMessage()) {
+                // Task was executed, mark as deleted but keep it.
+                $task->setDeleted(true);
+                $this->em->persist($task);
+            } else {
+                // Task was never executed.
+                $this->em->remove($task);
+            }
         }
 
         $this->em->flush();
@@ -88,15 +83,17 @@ class TaskService
     }
 
     /**
-     * @param \Symfony\Component\Console\Output\OutputInterface $output
+     * @param OutputInterface $output
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public function runTasks(OutputInterface $output): void
     {
         $due = new DateTime();
 
-        /** @var Task[] $tasks */
+        /**
+         * @var Task[] $tasks
+         */
         $tasks = $this->em->getRepository(Task::class)->findAllDue($due);
 
         foreach ($tasks as $task) {
@@ -106,22 +103,31 @@ class TaskService
                 continue;
             }
 
+            $taskUuid = $task->getUuid();
             $commandClass = $task->getCommand();
             $aggregateUuid = $task->getAggregateUuid();
             $payload = $task->getPayload();
 
             // Get current aggregate version.
-            /** @var EventStreamObject[]|null $lastEventOnAggregate */
+            /**
+             * @var EventStreamObject[]|null $lastEventOnAggregate
+             */
             $lastEventOnAggregate = $this->em->getRepository(EventStreamObject::class)->findBy(['uuid' => $aggregateUuid], ['version' => 'DESC'], 1);
             if (!empty($lastEventOnAggregate[0])) {
                 $onVersion = $lastEventOnAggregate[0]->getVersion();
+                $aggregateClass = $lastEventOnAggregate[0]->getAggregateClass();
             } else {
                 $onVersion = null;
+                $aggregateClass = null;
             }
 
-            if (null !== $onVersion) {
-                $command = new $commandClass(-1, null, $aggregateUuid, $onVersion, $payload);
-                $this->commandBus->dispatch($command);
+            if (null !== $onVersion && null !== $aggregateClass) {
+                $success = $this->commandBus->execute(
+                    $commandClass,
+                    $aggregateUuid,
+                    $payload,
+                    -1
+                );
 
                 // Save messages on task.
                 $messages = $this->messageBus->getMessages();
@@ -131,6 +137,14 @@ class TaskService
                 $this->em->flush();
 
                 $output->writeln('Executed task '.$commandClass.' on '.$aggregateUuid);
+
+                // Remove task from page.
+                if ($success && $aggregateClass === Page::class) {
+                    $command = new PageRemoveScheduleCommand(-1, null, $aggregateUuid, ($onVersion+1), [
+                        'scheduleUuid' => $taskUuid,
+                    ]);
+                    $this->commandBus->dispatch($command);
+                }
             }
         }
